+++ SistemaEstoque/Model/editarbtn.php
@@ -1,5 +1,5 @@
-<?php  
-include "../conexaophp.php";
+<?php
+include "../../conexaophp.php";
 
 session_start();
 
@@ -8,17 +8,10 @@ $enderecoeditar = $_POST["localdest"];
 $quantidadeeditar = $_POST["quantidade"];
 $produto = $_POST["produto"];
 
-// echo $nunotadest;
-// echo "<br>".$enderecoeditar;
-// echo "<br>".$quantidadeeditar;
-// echo "<br>".$produto;
-
 $tsql2 = "EXEC AD_STP_EDITARPRODUTONOTA_SISTEMAESTOQUE $nunotadest, $enderecoeditar, $quantidadeeditar, '$produto'";
 
-$stmt2 = sqlsrv_query( $conn, $tsql2);
+$stmt2 = sqlsrv_query($conn, $tsql2);
 
 $row = sqlsrv_fetch_array($stmt2, SQLSRV_FETCH_NUMERIC);
 $retorno = $row[0];
 echo $retorno;
-
-?>
