--- SistemaEstoque/css/insereestoque2.css
@@ -1,291 +1,0 @@
-body, html {
-	height: 100%;
-	width: 100%;
-	margin: 0;
-	background-color: #FFFFFD;
-}
-
-.container {
-	height: 100%;
-	width: 100%;
-	margin: 0;
-	background-color: #FFFFFD;
-	//position: absolute;
-	display: flex;
-}
-
-.infonota {
-	display: flex;
-	background-color: #F3FFFF;
-	width: 60%;
-	height: 25%;
-	//margin: 0;
-	//justify-content: center;
-	align-items: center;
-	//text-align: center;
-}
-
-.infonotadiv {
-	width: 100%;
-	//background-color: yellow;
-	height: 70%;
-}
-
-.infonotatext {
-	margin-left: 5%;
-	line-height: 20px;
-	//margin: auto;
-}
-
-.infonotatextprod {
-	margin: auto;	
-}
-
-.acoes {
-	background-color: #DAE1F5;
-	width: 38%;
-	height: 25%;
-	//margin: 0;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	text-align: center;
-}
-
-.botoes {
-	
-}
-
-#finalizar {
-	margin-left: auto;
-	margin-top: auto;
-	background-color: green;
-	color: #FFFFFD;
-	border-radius: 10%;
-	text-align: center;
-	align-items: center;
-	//display: inline-block;
-	width: 80px;
-	height: 30px;
-}
-
-.editarbtn {
-	height: 30px;
-	margin-left: auto;
-	margin-top: 10%;
-	border-radius: 10%;
-	background-color: #FFF2E6;
-	//display: inline-block;
-}
-
-.divergenciasbtn{
-	height: 30px;
-	margin-left: auto;
-	margin-top: 10%;
-	border-radius: 10%;
-	background-color: #FFF2E6;
-	//display: inline-block;
-}
-
-.item {
-	width: 98%;
-	height: 74%;
-	background-color: #EBF7F7;
-	margin: 0;
-	position: fixed;
-	bottom: 0;
-	justify-content: center;
-	//display: flex;
-	text-align: center;
-}
-
-.cxtexto {
-	margin-top: 8px;
-}
-
-#confirmar {
-	width: 50%;
-	height: 30px;
-	border-radius: 10%;
-	font-weight: bold;
-}
-
-.foto {
-	width: 100px;
-	//background-color: yellow;
-	margin-left: auto;
-	margin-right: auto;
-	margin-top: 5%;
-	margin-bottom: 5%;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	text-align: center;
-}
-
-.foto-d {
-	display: flex;
-	align-items: center; 
-	display: inline-flex; 
-	margin-top: 0; 
-	padding-top: 0
-}
-
-.fechar{
-	position: absolute;
-	top: -10px;
-	right: -10px;
-	width: 30px;
-	height: 30px;
-	border-radius: 50%;
-	border: 0;
-	background-color: red;
-	color: white;
-	font-size: 20px;
-	cursor: pointer;
-}
-
-.popupprodutos{
-	position: fixed;
-	top: 0; bottom: 0; 
-	left: 0; right:0;
-	margin: auto;
-	width: 86%;
-	height: 90%;
-	padding: 15px;
-	border: solid 1px #4c4d4f;
-	background: white;
-	display: none;
-	z-index: 10;
-	border-radius: 10px;
-}
-
-.popupdivergencias{
-	position: fixed;
-	top: 0; bottom: 0; 
-	left: 0; right:0;
-	margin: auto;
-	width: 86%;
-	height: 90%;
-	padding: 15px;
-	border: solid 1px #4c4d4f;
-	background: white;
-	display: none;
-	z-index: 10;
-	border-radius: 10px;
-}
-
-.tempprodutos{
-	position: fixed;
-	top: 0; bottom: 0; 
-	left: 0; right:0;
-	margin: auto;
-	width: 86%;
-	height: 90%;
-	padding: 15px;
-	border: solid 1px #4c4d4f;
-	background: white;
-	display: none;
-	z-index: 10;
-	border-radius: 10px;
-}
-
-.popupEditar{
-	position: fixed;
-	top: 0; bottom: 0; 
-	left: 0; right:0;
-	margin: auto;
-	width: 60%;
-	height: 38%;
-	padding: 10px;
-	border: solid 1px #4c4d4f;
-	background: white;
-	display: none;
-	z-index: 999;
-	border-radius: 10px;
-}
-
-.popupEditarTemp{
-	position: fixed;
-	top: 0; bottom: 0; 
-	left: 0; right:0;
-	margin: auto;
-	width: 60%;
-	height: 38%;
-	padding: 10px;
-	border: solid 1px #4c4d4f;
-	background: white;
-	display: none;
-	z-index: 999;
-	border-radius: 10px;
-}
-
-.popupInserirEndereco{
-	position: fixed;
-	top: 0; bottom: 0; 
-	left: 0; right:0;
-	margin: auto;
-	width: 60%;
-	height: 16%;
-	padding: 10px;
-	border: solid 1px #4c4d4f;
-	background: white;
-	display: none;
-	z-index: 999;
-	border-radius: 10px;
-}
-
-table {
-	border-collapse: collapse;
-}
-
-table th {
-	text-align: center;
-	background-color: #3a6070;
-	color: #FFF;
-	/*padding: 4px 30px 4px 8px;*/
-}
-
-table td {
-	border: 1px solid #e3e3e3;
-	padding: 4px 8px;
-}
-
-table tr:nth-child(odd) td {
-	background-color: #e7edf0;
-}
-
-.excluir {
-	background-color: red;
-	color: white;
-	border: 0;
-	border-radius: 25%;
-}
-
-.editar {
-	background-color: white;
-	border: 0;
-	border-radius: 25%;
-}
-
-#editarTempBtn {
-	display: none;
-}
-
-
-
-#loader{
-	position: fixed;
-	z-index: 99999;
-	left: 0px;
-	right: 0px;
-	width: 100%;
-	height: 100%;
-	background: 50% 50% no-repeat rgb(15 10 10/50%);
-}
-#loader img{
-	position: absolute;
-  	top: 30%;
-  	left: 50%;
-  	transform: translate(-50%, -50%);
-}
